Reference

PicardLefschetzIntegration.index

PicardLefschetzIntegration.parameters

PicardLefschetzIntegration.point

PicardLefschetzIntegration.thimble

PicardLefschetzIntegration.PL_integrate

PicardLefschetzIntegration.divide

PicardLefschetzIntegration.divide_rep

PicardLefschetzIntegration.flow!

PicardLefschetzIntegration.flow_point

PicardLefschetzIntegration.flow_points

PicardLefschetzIntegration.get_affine_map

PicardLefschetzIntegration.hypercube_triangulation_indices

PicardLefschetzIntegration.initialGrid

PicardLefschetzIntegration.integrateSimplex

PicardLefschetzIntegration.jacobian

PicardLefschetzIntegration.mapping

PicardLefschetzIntegration.normalize

PicardLefschetzIntegration.remove_inactive_points

PicardLefschetzIntegration.volume_simplex

PicardLefschetzIntegration.index — Type

index

The indices that span a simplex. coord::Vector{Int} The indices of the simplex refering to vertices in the vector points. active When the simplex is active, it is still part of the Picard Lefschetz flow.

PicardLefschetzIntegration.parameters — Type

parameters

Contains the parameters of the Picard Lefschetz integration method.

dim         The dimension of the integral.
δ           The maximal edge length of the simplices.
τ           Threshold on the real part of the exponent.
ϵ           The steps size of the downward flow.
N           The number of steps in the downward flow.
n           The order of the simplex quadrature integration scheme.

PicardLefschetzIntegration.point — Type

point

A point with two boolian attributes. The first determines whether the point is active. The second determines whether the point flows or stays fixed.

coord::Vector{Complex}  The coordinates of the point.
active                  When the point is active, it is still part of the Picard Lefschetz flow.
frozen                  When a point is fozen, it does not flow. This is particularly relevant for points on the boundary of the integration domain.

PicardLefschetzIntegration.thimble — Type

thimble

The thimble consisting of a set of points and a set of simplices.

PicardLefschetzIntegration.PL_integrate — Method

PL_integrate(S, thim::thimble, pars::parameters)

Evaluate the integral ∫exp(im * S(x))dx along the thimble.

PicardLefschetzIntegration.divide — Method

divide(thim, pars::parameters)

Subdivide the simplices that have an edge longer than the threshold pars.δ.

PicardLefschetzIntegration.divide_rep — Method

divide_rep(thim::thimble, pars::parameters)

Repeat subdivision till all edges of the simplices are shorter than par.δ.

PicardLefschetzIntegration.flow! — Method

flow!(S, thim::thimble, pars::parameters)

Flow the thimble pars.N steps forward in the downward flow.

PicardLefschetzIntegration.flow_point — Method

flow_point(S, p, pars::parameters)

Push a point one step in the downward flow.

PicardLefschetzIntegration.flow_points — Method

flow_points(S, thim::thimble, pars::parameters)

Flow the thimble one step in the downward flow.

PicardLefschetzIntegration.get_affine_map — Method

get_affine_map(vertices)

Affine map mapping a general simplex to a standard simplex.

PicardLefschetzIntegration.hypercube_triangulation_indices — Method

hypercube_triangulation_indices(dim::Int)

Triangulate a dim dimensional hypercube and output the simplices.

PicardLefschetzIntegration.initialGrid — Function

initialGrid(min, max, pars::parameters, frozen = false)

Initialize a cubic grid with the lower left corner min and the upper right corner max. When forzen = true, the boundary points do not flow.

PicardLefschetzIntegration.integrateSimplex — Method

integrateSimplex(f, vertices, X, W)

Integrae f over a simplex spanned by vertices with a simplex quadrature method with the points X and the weights W.

PicardLefschetzIntegration.jacobian — Method

jacobian(vertices)

Evaluate the Jacobian of the affine map

PicardLefschetzIntegration.mapping — Method

mapping(p, vertices)

Map the standard simplex to a point in the simplex spanned by vertices.

PicardLefschetzIntegration.normalize — Function

normalize(p, thres = 1.)

Normalize a vector when the norm exceeds the threshold.

PicardLefschetzIntegration.remove_inactive_points — Method

remove_inactive_points(thim::thimble)

Remove inactive points.

PicardLefschetzIntegration.volume_simplex — Method

volume_simplex(vertices)

The volume of a simplex.